+++ assets/javascripts/kitten/components/cards/project-creator-card.js
@@ -36,8 +36,9 @@ class ProjectCreatorCardComponent extends React.Component {
 
     return (
       <div className="k-ProjectCreatorCard__status">
-        <span className="k-ProjectCreatorCard__status--icon"
-              style={ statusStyles }>
+        <span
+          className="k-ProjectCreatorCard__status--icon"
+          style={ statusStyles }>
         </span>
         <div className="k-ProjectCreatorCard__status--text">
           { statusText }
@@ -49,15 +50,17 @@ class ProjectCreatorCardComponent extends React.Component {
   renderDescription() {
     return (
       <div className="k-ProjectCreatorCard__grid">
-        <Marger top="2"
-                bottom="1">
+        <Marger
+          top="2"
+          bottom="1">
           <div className="k-ProjectCreatorCard__grid--flex">
             { this.renderDate() }
             { this.renderStatus() }
           </div>
         </Marger>
-        <Marger top="1"
-                bottom="2">
+        <Marger
+          top="1"
+          bottom="2">
           <Paragraph margin={ false }>
             { this.props.children }
           </Paragraph>
@@ -77,17 +80,17 @@ class ProjectCreatorCardComponent extends React.Component {
          { ...target }>
         <Separator />
         <div className="k-ProjectCreatorCard__grid">
-          <Marger top="1.2"
-                  bottom="1.2"
-                  className="k-ProjectCreatorCard__grid--flex">
+          <Marger
+            top="1.2"
+            bottom="1.2"
+            className="k-ProjectCreatorCard__grid--flex">
             { this.props.linkText }
-            <div className="k-ProjectCreatorCard__link--withAnimation">
-              <ButtonIcon className={ classNames('k-ButtonIcon--withoutHover',
-                                                 'k-ButtonIcon--verticalArrow') }
-                          size="tiny">
-                <RightArrowIcon className="k-ButtonIcon__svg" />
-              </ButtonIcon>
-            </div>
+            <ButtonIcon
+              size="tiny"
+              withoutHover
+              verticalArrow>
+              <RightArrowIcon className="k-ButtonIcon__svg" />
+            </ButtonIcon>
           </Marger>
         </div>
       </a>