+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `Modal` component.
+- Fix: Add `withoutHover` and `verticalArrow` options on `ButtonIcon` component.
 - Fix: Add `ButtonIcon` animation on `ProjectCreatorCard`.
 - Feature: Add `k-u-align-center-block` utility to center blocks.